assets moved to 'jquery'
