+++ framework/jquery/YiiAsset.php
@@ -5,31 +5,28 @@
  * @license http://www.yiiframework.com/license/
  */
 
-namespace yii\validators;
+namespace yii\jquery;
 
 use yii\web\AssetBundle;
-use yii\web\YiiAsset;
 
 /**
- * This asset bundle provides the javascript files for client validation.
+ * This asset bundle provides the base JavaScript files for the Yii Framework.
  *
  * @author Qiang Xue <qiang.xue@gmail.com>
  * @since 2.0
  */
-class ValidationAsset extends AssetBundle
+class YiiAsset extends AssetBundle
 {
     /**
      * @inheritdoc
      */
-    public $sourcePath = '@yii/assets';
-
+    public $sourcePath = '@yii/jquery/assets';
     /**
      * @inheritdoc
      */
-    public $js = ['yii.validation.js'];
-
+    public $js = ['yii.js'];
     /**
      * @inheritdoc
      */
-    public $depends = [YiiAsset::class];
+    public $depends = [JqueryAsset::class];
 }
